refactor(tee): split sealed_secrets_api.rs god object into endpoint modules
tee/sealed_secrets_api.rs (619 LOC) -> tee/sealed_secrets_api/: mod.rs keeps
the module doc, imports, measurement-pinning config (require_pinned_/
release_routes_enabled), the shared PublicKeyResponse + enforce_release_gate,
and re-exports the tee nonce helpers reached via super::. keys.rs holds the
public-key + sealed-secret-injection endpoints; attestation.rs holds the
attestation challenge/response endpoints. Endpoint handlers stay pub and are
re-exported so operator_api routing still reaches them; submodules read the
gate/types as children of mod.rs (no widening needed). Pure code motion, no
behavior change; max module 181 LOC.
