+++ scripts/.god-object-baseline.txt
@@ -15,5 +15,4 @@
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
 627 ui/src/lib/hooks/useCreateDeploy.ts
 624 ui/src/lib/api/sandboxClient.ts
-619 sandbox-runtime/src/tee/sealed_secrets_api.rs
 606 sandbox-runtime/src/util.rs
